Add files via upload
These are the 5 Tasks given to me by #codsoft @codsoft internship in Python programming. 
These are Zip files for all of these which include a python file for the main project, Jupyter File for command line output, and a Folder for images or Icons,   for every task.
Please change the paths for the images/icons used in this repository as per your system for them to work well.